cohttp-eio(feat): request body, handler
This commit implements reading request body as either a String or a
chunked body. Chunked body is read as a streaming function.

The commit also introduces support for handler and middleware so that
we can create composable web applications.

Lastly, a request_loop handles exceptions and sends response to
clients as required.