cohttp-eio(feat): make Body.t reusable
This commit aims to make `'a Body.t` reusable across both request and
response. It therefore rename request.ml to body.ml and updates all
references to the old Request module as Body. This commit should pave
way to being able to write response body (whether string or chunked)
as we do when reading requests.