cohttp-eio: remove faraday
Removes faraday as the primary serializer for http response. We use
good old `sprintf` and Buffer.t as a replacement. Faraday was adding
unnecessary loop overhead - loop running and being yielded most of
the time it was run. Since HTTP 1.1 is mostly synchronous, a direct
style serialization is sufficient for our needs.

Updates eio to the latest version so that we can benchmark it.